+++ components/ButtonTheme.vue
@@ -23,7 +23,7 @@
                         slot="placeholder"
                         v-bind="background"
                         class="relative w-64 h-48 overflow-hidden rounded-lg"
-                    ></div>
+                    />
                 </LazyComponent>
             </button>
 
@@ -93,7 +93,7 @@ export default {
         const blocks = ref([]);
         const visible = ref(false);
         const rendered = ref(false);
-        const themeSettings = reactive(defaults({ scale: 0.5 }, cloneDeep(settings.value)));
+        const themeSettings = reactive({});
 
         const generateTokens = () =>
             buildCodeBlocks(
@@ -108,6 +108,12 @@ export default {
                     themeSettings.themeBackground = background;
                 }
             );
+
+        watch(
+            settings,
+            (values) => Object.assign(themeSettings, defaults({ scale: 0.5 }, cloneDeep(values))),
+            { immediate: true, deep: true }
+        );
 
         onMounted(() => {
             watch(visible, (visible) => {